requirements.txt
requests
python-dotenv
moviepy
pillow
numpy
guidance
# locked to this version because guidance needs to be updated.
openai==0.28.1
Flask
gevent